# Speedup DimShuffle and Reshape in C and Numba backends