… tests/tensor/test_shape.py