## TLDR: Make views cheap again!
### All backends
Remove the unused inplace option, less branching is always helpful

### Python backend
~~Implementation based on `as_strided` that mimicks the old (and now current again) C-backend approach. This is debatable since the Python implementation has historically been there more for readability than performance. However it's useful for fast compiling graphs not to suck?~~
`as_strided` is actually insanely slow in python, so I'm back to transpose+reshape. The removal of some checks and unnecessary `asarray` calls provide a small speedup, although this is obviously not critical.

### Numba backend
Use `as_strided` which simplifies the implementation by a ton with two benefits:
1. Much faster compilation / caching. The new benchmark runs on 1.6s (includes many runs) after caching, vs 4.6s. When checking a simple function directly I saw wall times going down from 1.2 vs 600ms (first time after starting the interpreter) and 300ms vs 100ms in subsequent runs. This seems to scale with the number of DimShuffles in the graph?
2. Slightly faster for contiguous arrays: the new benchmark roughly 1.1-1.3x faster

```python
import pytensor
import pytensor.tensor as pt
import numpy as np

x = pt.tensor("x", shape=(1, 2, 3))
y = x.dimshuffle(2, "x", 1, "x")

out = pytensor.function([pytensor.In(x, borrow=True)], pytensor.Out(y, borrow=True), mode="NUMBA")
out.trust_input = True

x_test = np.zeros((1, 2, 3))
out(x_test)
%timeit out(x_test)
```

The example with non-contiguous arrays in https://github.com/pymc-devs/pytensor/issues/1111 has no penalty now, which is a 10,000x speedup (example is exaggerated with a very large array)


### C backend
This was the original goal of this PR

There were two regressions caused by e593b0ac57a0d56d4f6ffdd08d52c3be78ebf961 and 223ee1548574b6bb8e73611ed605a97e29f13e7b.

The DimShuffle changes introduce a `PyArray_IntpConverter` to compute a new shape and used two calls to numpy C-functions `PyArray_Transpose` and `PyArray_NewShape`. I suspect the slowdown comes mostly from the introduction of `PyArray_IntPConverter` but I couldn't find anything wrong with the simpler logic from Theano times, other than...

The bug that motivated the changes had to do with a useless second pass on the dimensions calculations:
```C
    /* set the strides of the broadcasted dimensions.
     * This algorithm is from numpy: PyArray_Newshape() in
     * cvs/numpy/numpy/core/src/multiarraymodule.c */
    if (nd_out > 0) {
        if (strides[nd_out - 1] == 0)
            strides[nd_out - 1] = PyArray_DESCR(basename)->elsize;
        for (npy_intp i = nd_out - 2; i > -1; --i) {
            if (strides[i] == 0)
                strides[i] = strides[i + 1] * dimensions[i + 1];
        }
    }
```
Which is baffling in that DimShuffle is not doing any broadcasting behavior, othen than expand_dims, which the first pass already handles correctly. Removing this loop fixes the original bug. 

The Reshape changes were simpler, they introduced a generic `PyArray_IntpConverter` to convert the shape numpy vector into a simple C-array for the Reshape operation.  

The concern about strides in the commit message doesn't hold, because they were being used directly. I added a direct tests for odd strides just in case.

In the long term Reshape should take the separate shape entries as scalars, instead of a single vector. The user never defines a graph with a vector anyway, but with isolated entries, so the whole packing and unpacking is just overhead. This is tracked #881.

Profiling a simple function suggests the previous changes caused a 8.6-11x slowdown per Op call for the DimShuffle operation and a 5.4x slowdown for the Reshape operation. The new benchmarks detects only a 3.6x and 2.3x slowdown, respectively, due to the PyTensor call overhead.


```
## Op
### DimShuffle: (8.6x slowdown)
Before: 1.74e-06s per call (class)
After: 2.02e-07s

### Reshape: (5.4x slowdown)
Before: 1.50e-06s per call (class)
After: 2.78e-07s
```