@jessegrabowski 2025-02-23
Running the test suit locally, I see that the `test_solve_triangular_grad` test is also failing, but only when `lower=False`, which is the mode that is used in the cholesky gradient, so the bug might be there?