@ricardoV94 2025-02-23
Found the bug: https://github.com/pymc-devs/pytensor/issues/1233